+++ src/lib.rs
@@ -18,7 +18,7 @@
 //!
 //! ## Example
 //! ```
-//! use fixnum::{FixedPoint, typenum::U9, ops::{CheckedAdd, RoundingMul, RoundMode::*}};
+//! use fixnum::{FixedPoint, typenum::U9, ops::{CheckedAdd, RoundingMul, RoundMode::*, Zero}};
 //!
 //! /// Signed fixed point amount over 64 bits, 9 decimal places.
 //! ///
@@ -30,12 +30,14 @@
 //! ///           = 5e-10
 //! type Amount = FixedPoint<i64, U9>;
 //!
-//! fn amount(s: &str) -> Amount { s.parse().unwrap() }
-//!
 //! # fn main() -> Result<(), Box<dyn std::error::Error>> {
-//! assert_eq!(amount("0.1").cadd(amount("0.2"))?, amount("0.3"));
-//! let expences: Amount = amount("0.000000001");
-//! assert_eq!(expences.rmul(expences, Floor)?, amount("0.0"));
+//! let a: Amount = "0.1".parse()?;
+//! let b: Amount = "0.2".parse()?;
+//! assert_eq!(a.cadd(b)?, "0.3".parse()?);
+//!
+//! let expences: Amount = "0.000000001".parse()?;
+//! // 1e-9 * (Floor) 1e-9 = 0
+//! assert_eq!(expences.rmul(expences, Floor)?, Amount::ZERO);
 //! // 1e-9 * (Ceil) 1e-9 = 1e-9
 //! assert_eq!(expences.rmul(expences, Ceil)?, expences);
 //! # Ok(()) }
@@ -48,9 +50,14 @@
 //! | [`cadd`][cadd] | `let result: Result<FixedPoint, ArithmeticError> = a.cadd(b)` | Checked addition. Returns `Err` on overflow. |
 //! | [`csub`][csub] | `let result: Result<FixedPoint, ArithmeticError> = a.csub(b)` | Checked subtraction. Returns `Err` on overflow. |
 //! | [`cmul`][cmul] | `let result: Result<FixedPoint, ArithmeticError> = a.cmul(b)` | Checked multiplication. Returns `Err` on overflow. This is multiplication without rounding, hence it's available only when at least one operand is integer. |
-//! | [`rmul`][rmul] | `let result: Result<FixedPoint, ArithmeticError> = a.rmul(b, RoundMode::Ceil)` | Checked rounded multiplication. Returns `Err` on overflow. Because of provided [`RoundMode`][RoundMode] it's possible across the [`FixedPoint`][FixedPoint] values. |
-//! | [`rdiv`][rdiv] | `let result: Result<FixedPoint, ArithmeticError> = a.rdiv(b, RoundMode::Floor)` | Checked rounded division. Returns `Err` on overflow. Because of provided [`RoundMode`][RoundMode] it's possible across the [`FixedPoint`][FixedPoint] values. |
+//! | [`rmul`][rmul] | `let result: Result<FixedPoint, ArithmeticError> = a.rmul(b, RoundMode::Ceil)` | Checked rounding multiplication. Returns `Err` on overflow. Because of provided [`RoundMode`][RoundMode] it's possible across the [`FixedPoint`][FixedPoint] values. |
+//! | [`rdiv`][rdiv] | `let result: Result<FixedPoint, ArithmeticError> = a.rdiv(b, RoundMode::Floor)` | Checked rounding division. Returns `Err` on overflow. Because of provided [`RoundMode`][RoundMode] it's possible across the [`FixedPoint`][FixedPoint] values. |
 //! | [`cneg`][cneg] | `let result: Result<FixedPoint, ArithmeticError> = a.cneg()` | Checked negation. Returns `Err` on overflow (you can't negate [`MIN` value][MIN]). |
+//! | [`integral`][integral] | `let y: {integer} = x.integral(RoundMode::Floor)` | Takes [rounded][RoundMode] integral part of the number. |
+//! | [`saturating_add`][saturating_add] | `let z: FixedPoint = x.saturating_add(y)` | Saturating addition |
+//! | [`saturating_sub`][saturating_sub] | `let z: FixedPoint = x.saturating_sub(y)` | Saturating subtraction |
+//! | [`saturating_mul`][saturating_mul] | `let z: FixedPoint = x.saturating_mul(y)` | Saturating multiplication. This is multiplication without rounding, hence it's available only when at least one operand is integer. |
+//! | [`saturating_rmul`][saturating_rmul] | `let z: FixedPoint = x.saturating_rmul(y, RoundMode::Floor)` | Saturating rounding multiplication |
 //!
 //! ## Implementing wrapper types.
 //! It's possible to restrict the domain in order to reduce chance of mistakes.
@@ -99,11 +106,16 @@
 //! [cmul]: ./ops/trait.CheckedMul.html#tymethod.cmul
 //! [fixnum]: ./macro.fixnum.html
 //! [FixedPoint]: ./struct.FixedPoint.html
-//! [MIN]: ./ops/trait.Numeric.html#associatedconstant.MIN
+//! [integral]: ./struct.FixedPoint.html#method.integral
+//! [MIN]: ./ops/trait.Bounded.html#associatedconstant.MIN
 //! [parity_scale_codec]: https://docs.rs/parity-scale-codec
 //! [rdiv]: ./ops/trait.RoundingDiv.html#tymethod.rdiv
 //! [rmul]: ./ops/trait.RoundingMul.html#tymethod.rmul
 //! [RoundMode]: ./ops/enum.RoundMode.html
+//! [saturating_add]: ./ops/trait.CheckedAdd.html#tymethod.saturating_add
+//! [saturating_mul]: ./ops/trait.CheckedMul.html#tymethod.saturating_mul
+//! [saturating_rmul]: ./ops/trait.RoundingMul.html#tymethod.saturating_rmul
+//! [saturating_sub]: ./ops/trait.CheckedSub.html#tymethod.saturating_sub
 
 #![warn(rust_2018_idioms)]
 #![cfg_attr(not(feature = "std"), no_std)]
@@ -118,7 +130,7 @@ use typenum::Unsigned;
 #[cfg(feature = "i128")]
 use crate::i256::I256;
 use crate::ops::{
-    CheckedAdd, CheckedMul, CheckedSub, Numeric, RoundMode, RoundingDiv, RoundingMul,
+    Bounded, CheckedAdd, CheckedMul, CheckedSub, One, RoundMode, RoundingDiv, RoundingMul, Zero,
 };
 pub use typenum;
 
@@ -199,9 +211,15 @@ macro_rules! impl_fixed_point {
             const COEF_PROMOTED: $promotion = $convert(Self::COEF) as _;
         }
 
-        impl<P: Precision> Numeric for FixedPoint<$layout, P> {
+        impl<P: Precision> Zero for FixedPoint<$layout, P> {
             const ZERO: Self = Self::from_bits(0);
+        }
+
+        impl<P: Precision> One for FixedPoint<$layout, P> {
             const ONE: Self = Self::from_bits(Self::COEF);
+        }
+
+        impl<P: Precision> Bounded for FixedPoint<$layout, P> {
             const MIN: Self = Self::from_bits($layout::MIN);
             const MAX: Self = Self::from_bits($layout::MAX);
         }
@@ -314,6 +332,11 @@ macro_rules! impl_fixed_point {
                     .map(Self::from_bits)
                     .ok_or(ArithmeticError::Overflow)
             }
+
+            #[inline]
+            fn saturating_add(self, rhs: Self) -> Self::Output {
+                Self::Output::from_bits(self.inner.saturating_add(rhs.inner))
+            }
         }
 
         impl<P: Precision> CheckedSub for FixedPoint<$layout, P> {
@@ -326,6 +349,11 @@ macro_rules! impl_fixed_point {
                     .checked_sub(rhs.inner)
                     .map(Self::from_bits)
                     .ok_or(ArithmeticError::Overflow)
+            }
+
+            #[inline]
+            fn saturating_sub(self, rhs: Self) -> Self::Output {
+                Self::Output::from_bits(self.inner.saturating_sub(rhs.inner))
             }
         }
 
@@ -340,6 +368,11 @@ macro_rules! impl_fixed_point {
                     .map(Self::from_bits)
                     .ok_or(ArithmeticError::Overflow)
             }
+
+            #[inline]
+            fn saturating_mul(self, rhs: $layout) -> Self::Output {
+                Self::Output::from_bits(self.inner.saturating_mul(rhs))
+            }
         }
 
         impl<P: Precision> CheckedMul<FixedPoint<$layout, P>> for $layout {
@@ -349,6 +382,11 @@ macro_rules! impl_fixed_point {
             #[inline]
             fn cmul(self, rhs: FixedPoint<$layout, P>) -> Result<FixedPoint<$layout, P>> {
                 rhs.cmul(self)
+            }
+
+            #[inline]
+            fn saturating_mul(self, rhs: FixedPoint<$layout, P>) -> Self::Output {
+                Self::Output::from_bits(self.saturating_mul(rhs.inner))
             }
         }
 
@@ -360,7 +398,7 @@ macro_rules! impl_fixed_point {
 
             /// Checked negation. Returns `Err` on overflow (you can't negate [`MIN` value][MIN]).
             ///
-            /// [MIN]: ./ops/trait.Numeric.html#associatedconstant.MIN
+            /// [MIN]: ./ops/trait.Bounded.html#associatedconstant.MIN
             #[inline]
             pub fn cneg(self) -> Result<FixedPoint<$layout, P>> {
                 self.inner
@@ -383,6 +421,25 @@ macro_rules! impl_fixed_point {
                 }
             }
 
+            /// Takes [rounded][RoundMode] integral part of the number.
+            ///
+            /// ```
+            /// use fixnum::{FixedPoint, typenum::U9, ops::RoundMode::*};
+            ///
+            /// type Amount = FixedPoint<i64, U9>;
+            ///
+            /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+            /// let a: Amount = "8273.519".parse()?;
+            /// assert_eq!(a.integral(Floor), 8273);
+            /// assert_eq!(a.integral(Ceil), 8274);
+            ///
+            /// let a: Amount = "-8273.519".parse()?;
+            /// assert_eq!(a.integral(Floor), -8274);
+            /// assert_eq!(a.integral(Ceil), -8273);
+            /// # Ok(()) }
+            /// ```
+            ///
+            /// [RoundMode]: ./ops/enum.RoundMode.html
             #[inline]
             pub fn integral(self, mode: RoundMode) -> $layout {
                 let sign = self.inner.signum();